staging before changing settings

+++ lib/gui/subWindowSettings.py
@@ -29,29 +29,27 @@ class Settings:
         self.app.stopSubWindow()
 
     def press(self, button):
-
         if button == "Update":
-
-            oldPath = configPytest.get('pytest', 'testpaths')
-            new= self.app.getEntry("testPathField")
-            oldAddopts = configPytest.get('pytest', 'addopts')
-            newAddopts = self.app.getEntry("addoptsField")
-
-            if oldPath != new:
-                configPytest.set('pytest', 'testpaths', new)
-                with open('pytest.ini', 'w') as f:
-                    configPytest.write(f)
-                self.app.setLabel("CurrentTestPath", f" Current testPath : {configPytest.get('pytest','testpaths')}")
-                self.app.setEntryDefault("testPathField",  configPytest.get('pytest','testpaths'))
-                #self.app.hideSubWindow("Settings")
-                #self.app.showSubWindow("Settings")
-
-            if oldAddopts != newAddopts:
-                configPytest.set('pytest', 'addopts', newAddopts)
-                with open('pytest.ini', 'w') as f:
-                    configPytest.write(f)
-                self.app.setLabel("CurrentAddopts", f" Current addopts : {configPytest.get('pytest', 'addopts')}")
-                self.app.setEntryDefault("addoptsField", configPytest.get('pytest', 'addopts'))
-
+            if self.app.yesNoBox("YES&NO", "Are you Sure") == True:
+                oldPath = configPytest.get('pytest', 'testpaths')
+                new= self.app.getEntry("testPathField")
+                oldAddopts = configPytest.get('pytest', 'addopts')
+                newAddopts = self.app.getEntry("addoptsField")
+                if oldPath != new:
+                    configPytest.set('pytest', 'testpaths', new)
+                    with open('pytest.ini', 'w') as f:
+                        configPytest.write(f)
+                    self.app.setLabel("CurrentTestPath", f" Current testPath : {configPytest.get('pytest','testpaths')}")
+                    self.app.setEntryDefault("testPathField",  configPytest.get('pytest','testpaths'))
+                    #self.app.hideSubWindow("Settings")
+                    #self.app.showSubWindow("Settings")
+                if oldAddopts != newAddopts:
+                    configPytest.set('pytest', 'addopts', newAddopts)
+                    with open('pytest.ini', 'w') as f:
+                        configPytest.write(f)
+                    self.app.setLabel("CurrentAddopts", f" Current addopts : {configPytest.get('pytest', 'addopts')}")
+                    self.app.setEntryDefault("addoptsField", configPytest.get('pytest', 'addopts'))
+            else:
+                pass
         elif button == "Close":
             self.app.hideSubWindow("Settings")
